# doc/api.yaml
openapi: 3.0.0
paths:
...
  /favs:
    {
      get:
        {
          operationId: FavsController_getFavs,
          parameters: [],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Favourites' } },
                    },
                },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
    }
  '/favs/track/{id}':
    {
      post:
        {
          operationId: FavsController_addTrackToFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '201': { description: 'Successfully added to favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '422': { description: 'Track does not exist' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: FavsController_deleteTrackFromFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '204': { description: 'Successfully deleted from favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Track is not favourite' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
    }
  '/favs/artist/{id}':
    {
      post:
        {
          operationId: FavsController_addArtistToFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '201': { description: 'Successfully added to favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '422': { description: 'Artist does not exist' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: FavsController_deleteArtistFromFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '204': { description: 'Successfully deleted from favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Artist is not favourite' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
    }
  '/favs/album/{id}':
    {
      post:
        {
          operationId: FavsController_addAlbumToFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '201': { description: 'Successfully added to favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '422': { description: 'Album does not exist' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: FavsController_deleteAlbumFromFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '204': { description: 'Successfully deleted from favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Album is not favourite' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
    }
...
  '/album/{id}':
    {
      get:
        {
          operationId: AlbumsController_getAlbumById,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Album' },
                            },
                        },
                    },
                },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '404': { description: 'Album not found' },
            },
          tags: [albums],
          security: [{ bearer: [] }],
        },
      put:
        {
          operationId: AlbumsController_updateAlbum,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/AlbumDto' } },
                },
            },
          responses:
            {
              '200':
                {
                  description: 'Successfully updated',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Album' },
                            },
                        },
                    },
                },
              '400': { description: 'Invalid id or dto' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Album not found' },
            },
          tags: [albums],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: AlbumsController_deleteAlbum,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '204': { description: 'Successfully deleted' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Album not found' },
            },
          tags: [albums],
          security: [{ bearer: [] }],
        },
    }
...
  '/artist/{id}':
    {
      get:
        {
          operationId: ArtistsController_getArtistById,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Artist' } },
                    },
                },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '404': { description: 'Artist not found' },
            },
          tags: [artists],
          security: [{ bearer: [] }],
        },
      put:
        {
          operationId: ArtistsController_updateArtist,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/ArtistDto' } },
                },
            },
          responses:
            {
              '200':
                {
                  description: 'Successfully updated',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Artist' } },
                    },
                },
              '400': { description: 'Invalid id or dto' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Artist not found' },
            },
          tags: [artists],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: ArtistsController_deleteArtist,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '204': { description: 'Successfully deleted' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Artist not found' },
            },
          tags: [artists],
          security: [{ bearer: [] }],
        },
    }
...
  '/track/{id}':
    {
      get:
        {
          operationId: TracksController_getTrackById,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Track' } },
                    },
                },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '404': { description: 'Track not found' },
            },
          tags: [tracks],
          security: [{ bearer: [] }],
        },
      put:
        {
          operationId: TracksController_updateTrack,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/TrackDto' } },
                },
            },
          responses:
            {
              '200':
                {
                  description: 'Successfully updated',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Track' } },
                    },
                },
              '400': { description: 'Invalid id or dto' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Track not found' },
            },
          tags: [tracks],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: TracksController_deleteTrack,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '204': { description: 'Successfully deleted' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Album not found' },
            },
          tags: [tracks],
          security: [{ bearer: [] }],
        },
    }
...
info:
  title: 'Home Library'
  description: 'Home music library service'
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes: { bearer: { scheme: bearer, bearerFormat: JWT, type: http } }
  schemas:
...
      FindFavsDto:
        {
          type: object,
          properties: { userId: { type: string } },
          required: [userId],
        },
      Favourites:
        {
          type: object,
          properties:
            {
              id: { type: string },
              tracks: { type: array, items: { type: string } },
              artists: { type: array, items: { type: string } },
              albums: { type: array, items: { type: string } },
            },
          required: [id, tracks, artists, albums],
        },
      Album:
        {
          type: object,
          properties:
            {
              id: { type: string },
              name: { type: string },
              year: { type: number },
              artistId: { type: string },
            },
          required: [id, name, year, artistId],
        },
      AlbumDto:
        {
          type: object,
          properties:
            {
              name: { type: string },
              year: { type: number },
              artistId: { type: string },
            },
          required: [name, year],
        },
      Artist:
        {
          type: object,
          properties:
            {
              id: { type: string },
              name: { type: string },
              grammy: { type: boolean },
            },
          required: [id, name, grammy],
        },
      ArtistDto:
        {
          type: object,
          properties: { name: { type: string }, grammy: { type: boolean } },
          required: [name, grammy],
        },
      Track:
        {
          type: object,
          properties:
            {
              id: { type: string },
              name: { type: string },
              duration: { type: number },
              artistId: { type: string },
              albumId: { type: string },
            },
          required: [id, name, duration, artistId, albumId],
        },
      TrackDto:
        {
          type: object,
          properties:
            {
              name: { type: string },
              artistId: { type: string },
              albumId: { type: string },
              duration: { type: number },
            },
          required: [name, duration],
        },
...